Split bootstrap-list up into multiple files and create new ping stats transport for when we have a connection
