+++ packages/gateway-client/src/worker/p2p-client/bootstrap/index.ts
@@ -1,0 +1,1 @@
+export { BootstrapList } from './bootstrap-list';
